Players can no longer make a move that puts their king in jeopardy. Checkmate, and stalemates are now detected.
To get this to work I changed the structure of the Piece class. Now Piece is an abstract base class with child classes Pawn, Knight, Bishop, Rook, Queen, and King. I implemented a property on each Piece class called piece_type which returns the string "pawn", "knight", "bishop", "rook", "queen", and "king" depending on the class type of the piece. Each piece now reports what they believe to be their legal moves. This came in handy with invalidating moves based on whether or not the king was put into check. Pushing a move to the board no longer changes the state of the grid, now it just creates a new board state. To change the current board state I made a setter for grid, this is only called in the game class to ensure that the current board state remains without side affects until absolutely necessary.